Overworld API
I'm not completely happy with this. I know that an OverworldEntity needs
to map action sets with a trigger (such as stepping on, entering vision,
or interacting with), but I'm not sure the best way to model this
trigger mechanic. Obviously if triggers aren't something people want,
they can just key on null, and do whatever they want, essentially
ignoring the trigger functionality.